+++ core/anet-core/src/dqn_agent.cpp
@@ -149,18 +149,18 @@ public:
         map_["unstable_ema"] = vars.unstable_ema;
     }
 
-    virtual std::optional<float> GetScalar(const std::string& key, int index) const override {
+    virtual std::optional<float> GetScalar(const std::string& key, int64_t index) const override {
         auto itr = map_.find(key);
         if (itr == map_.end()) {
             return std::nullopt;
         }
         return itr->second;
     }
-    virtual std::optional<torch::Tensor> GetTensor(const std::string& key, int index) const override {
+    virtual std::optional<torch::Tensor> GetTensor(const std::string& key, int64_t index) const override {
         if (key == "max_q") return max_q_;
         return std::nullopt;
     }
-    virtual std::optional<std::vector<torch::Tensor>> GetTensorVector(const std::string& key, int index) const override {
+    virtual std::optional<std::vector<torch::Tensor>> GetTensorVector(const std::string& key, int64_t index) const override {
         return std::nullopt;
     }
 };
@@ -238,7 +238,7 @@ std::optional<anet::TensorFunction> DQNAgent::GetTensorFunction(const std::strin
     return std::nullopt;
 }
 
-std::optional<float> DQNAgent::GetScalar(const std::string& key, int index) const
+std::optional<float> DQNAgent::GetScalar(const std::string& key, int64_t index) const
 {
     if (key.find("replaybuffer.") == 0) {
         std::shared_lock<std::shared_mutex> lock(*mutex_);
@@ -256,7 +256,7 @@ std::optional<float> DQNAgent::GetScalar(const std::string& key, int index) cons
     return std::nullopt;
 }
 
-std::optional<torch::Tensor> DQNAgent::GetTensor(const std::string& key, int index) const
+std::optional<torch::Tensor> DQNAgent::GetTensor(const std::string& key, int64_t index) const
 {
     if (key.find("replaybuffer.") == 0)
         return replay_buffer_->GetTensor(key);
@@ -264,7 +264,7 @@ std::optional<torch::Tensor> DQNAgent::GetTensor(const std::string& key, int ind
     return std::nullopt;
 }
 
-std::optional<std::vector<torch::Tensor>> DQNAgent::GetTensorVector(const std::string& key, int index) const
+std::optional<std::vector<torch::Tensor>> DQNAgent::GetTensorVector(const std::string& key, int64_t index) const
 {
     if (key.find("replaybuffer.") == 0)
         return replay_buffer_->GetTensorVector(key);
