+++ core/anet-core/src/env.cpp
@@ -172,21 +172,21 @@ std::shared_ptr<DiscreteBatchEnvBase::StepResult> DiscreteBatchEnvBase::getStepR
 }
 
 
-std::optional<float> DiscreteBatchEnvBase::GetScalar(const std::string& key, int index) const
+std::optional<float> DiscreteBatchEnvBase::GetScalar(const std::string& key, int64_t index) const
 {
     ANET_ASSERT(index >= 0 && index < envs_.size());
     return envs_[index]->GetScalar(key, -1);
     /// @todo vec対応
 }
 
-std::optional<torch::Tensor> DiscreteBatchEnvBase::GetTensor(const std::string& key, int index) const
+std::optional<torch::Tensor> DiscreteBatchEnvBase::GetTensor(const std::string& key, int64_t index) const
 {
     ANET_ASSERT(index >= 0 && index < envs_.size());
     return envs_[index]->GetTensor(key, -1);
     /// @todo vec対応
 }
 
-std::optional<std::vector<torch::Tensor>> DiscreteBatchEnvBase::GetTensorVector(const std::string& key, int index) const
+std::optional<std::vector<torch::Tensor>> DiscreteBatchEnvBase::GetTensorVector(const std::string& key, int64_t index) const
 {
     ANET_ASSERT(index >= 0 && index < envs_.size());
     return envs_[index]->GetTensorVector(key, -1);
